Create expectation package (#74)
* Create expectation package

The purpose of this PR is to get ride of direct Kubernetes dependency. Expectation is being used in JobController to tell the number of pods/services they expect. This is the only library we reference in common code base. It makes sense to copy to our code base.

In the future, Instead of “import k8s.io/kubernetes/pkg/controller”,  User need to “import github.com/kubeflow/common/pkg/controller.v1/expectation" and use `expectation.NewControllerExpectations` instead.

Signed-off-by: Jiaxin Shan <seedjeffwan@gmail.com>

* Fix indirect usage of prometheus

* Use linter_config in CI